fix vercel deployment error - "authOptions" is not a valid Route export field
